[refactor] Tier agora tem Dao próprio. Nenhuma lógica alterada
